refactor: drop Constraints ZST shell, add enum Phase
- inline the EvalMode dispatch into PackContext::evaluate; move EvalMode/
  EvalOutput to context::eval and delete src/constraints/ (drops the
  molpack::constraints public path -> molpack::context)
- replace the bare 'phase == ntype' sentinel with enum Phase { PerType(usize),
  AllTypes }; pack()'s outer loop yields Phase values

Behavior-preserving: examples_batch parity green, 120 pytest pass.
