--- src/constraints/mod.rs
@@ -1,5 +1,0 @@
-//! Constraints container layer.
-
-mod container;
-
-pub use container::{Constraints, EvalMode, EvalOutput};
