+++ src/context/mod.rs
@@ -1,10 +1,12 @@
 //! Context layer for packmol-aligned packing runtime.
 
+pub mod eval;
 pub mod model;
 pub mod pack_context;
 pub mod state;
 pub mod work_buffers;
 
+pub use eval::{EvalMode, EvalOutput};
 pub use model::ModelData;
 pub use pack_context::{ATOM_FLAG_FIXED, ATOM_FLAG_SHORT, AtomProps, NONE_IDX, PackContext};
 pub use state::{RuntimeState, RuntimeStateMut};
